%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Color.PLOT.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{display.Color}
\alias{display.Color}
\alias{Color.display}
\alias{display.Color}
\alias{display,Color-method}


\title{Displays the current set of colors}

\usage{\method{display}{Color}(this, gridColor=NA, ..., label=c("none", "index", "rgb", "space"), labelColor=NULL, labelScale=0.6, labelRotate=0)}

\description{
 Displays the current set of colors.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Color}}.
}

\keyword{color}
\keyword{internal}
\keyword{methods}
