%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  OneDimensionalColor.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getColorSpace.OneDimensionalColor}
\alias{getColorSpace.OneDimensionalColor}
\alias{OneDimensionalColor.getColorSpace}
\alias{getColorSpace.OneDimensionalColor}
\alias{getColorSpace,OneDimensionalColor-method}


\title{Gets the internal representation of the colors}

\usage{\method{getColorSpace}{OneDimensionalColor}(this, ...)}

\description{
 Gets the internal representation of the colors.
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns the internal represenation of the colors as a column \code{\link[base]{vector}}.
 Most often these values are in the range [0,1], but that is not
 enforced.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{color}

\seealso{
  For more information see \code{\link{OneDimensionalColor}}.
}
\keyword{internal}
\keyword{methods}
