%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  WavelengthColor.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFrequency.WavelengthColor}
\alias{getFrequency.WavelengthColor}
\alias{WavelengthColor.getFrequency}
\alias{getFrequency.WavelengthColor}
\alias{getFrequency,WavelengthColor-method}


\title{Gets the frequences (in THz) of the colors}

\usage{\method{getFrequency}{WavelengthColor}(this, ...)}

\description{
 Gets the frequences (in THz) of the colors.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.colors:getWavelength.WavelengthColor]{*getWavelength}()}.
  For more information see \code{\link{WavelengthColor}}.
}

\keyword{color}
\keyword{internal}
\keyword{methods}
