%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  WavelengthColor.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getWavelength.WavelengthColor}
\alias{getWavelength.WavelengthColor}
\alias{WavelengthColor.getWavelength}
\alias{getWavelength.WavelengthColor}
\alias{getWavelength,WavelengthColor-method}


\title{Gets the wavelengths (in nm) of the colors}

\usage{\method{getWavelength}{WavelengthColor}(this, ...)}

\description{
 Gets the wavelengths (in nm) of the colors.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.colors:getFrequency.WavelengthColor]{*getFrequency}()}.
  For more information see \code{\link{WavelengthColor}}.
}

\keyword{color}
\keyword{internal}
\keyword{methods}
