%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Color.PLOT.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{image.Color}
\alias{image.Color}
\alias{Color.image}
\alias{image.Color}
\alias{image,Color-method}


\title{Creates an image of the current set of colors}

\usage{\method{image}{Color}(x, gridColor="gray", gridLty=3, gridLwd=1, axes=FALSE, main=NULL, ...)}

\arguments{
 \item{gridColor}{Color used for the grid.}
 \item{gridLty}{Line type (an \code{\link[base]{integer}}) for the grid lines.}
 \item{gridLwd}{Line width (a \code{\link[base]{numeric}}) of the grid lines.}
 \item{axes}{If \code{\link[base:logical]{TRUE}}, axes are shown, otherwise not.}
 \item{main}{Main title (a \code{\link[base]{character}} string).}
 \item{...}{Other arguments for \code{\link[graphics]{image}}.}
}

\description{
 Creates an image of the current set of colors.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Color}}.
}

\keyword{color}
\keyword{internal}
\keyword{methods}
