%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Color.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{length.Color}
\alias{length.Color}
\alias{Color.length}
\alias{length.Color}
\alias{length,Color-method}


\title{Gets the number of colors}

\usage{\method{length}{Color}(x)}

\description{
 Gets the number of colors.
}

\value{
 Returns the number of colors as an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{color}

\seealso{
  For more information see \code{\link{Color}}.
}
\keyword{internal}
\keyword{methods}
