%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Color.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{palette.Color}
\alias{palette.Color}
\alias{Color.palette}
\alias{palette.Color}
\alias{palette,Color-method}


\title{Use these colors as the color palette in all graphics}

\usage{\method{palette}{Color}(this, ...)}

\description{
 Use these colors as the color palette in all graphics.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{color}

\seealso{
  \code{\link[grDevices]{palette}}.
  For more information see \code{\link{Color}}.
}
\keyword{internal}
\keyword{methods}
