%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{R.colors-package}
\alias{R.colors-package}
\alias{R.colors}
\docType{package}

\title{Package R.colors}


 \description{
   Contains color and color filter classes.
 }

 \section{Installation}{
   To install this package, see \url{http://www.braju.com/R/}.
   Required packages are installed in the same way.
 }

 \section{Dependancies and other requirements}{
   This package depends on the \pkg{R.oo} and the \pkg{R.basic} packages.
 
   In addition, to use the \code{\link{HclColor}} class, the \pkg{colorspace}
   package must be installed. To use the \code{\link{DichromatColorFilter}} the
   \pkg{dichromat} package must be installed.
 }

 \section{Further readings}{
   Not available.
 }

 \section{How to cite this package}{
   Not available.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{package}
