%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  WavelengthColor.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{range.WavelengthColor}
\alias{range.WavelengthColor}
\alias{WavelengthColor.range}
\alias{range.WavelengthColor}
\alias{range,WavelengthColor-method}


\title{Gets the (wavelength) range of the colors}

\usage{\method{range}{WavelengthColor}(this, na.rm=TRUE, ...)}

\description{
 Gets the (wavelength) range of the colors.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.colors:getWavelength.WavelengthColor]{*getWavelength}()}.
  For more information see \code{\link{WavelengthColor}}.
}

\keyword{color}
\keyword{internal}
\keyword{methods}
