%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Color.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{seq.Color}
\alias{seq.Color}
\alias{Color.seq}
\alias{seq.Color}
\alias{seq,Color-method}


\title{Gets the indices of all colors}

\usage{\method{seq}{Color}(this, ...)}

\description{
 Gets the indices of all colors. This makes it easy to iterate over all colors (also when
 there are no colors).
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} of the indices of this vector.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{color}

\seealso{
  For more information see \code{\link{Color}}.
}
\keyword{internal}
\keyword{methods}
