%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RgbColor.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setBlue.RgbColor}
\alias{setBlue.RgbColor}
\alias{RgbColor.setBlue}
\alias{setBlue.RgbColor}
\alias{setBlue,RgbColor-method}


\title{Sets the color values in the blue channel}

\usage{\method{setBlue}{RgbColor}(this, ...)}

\description{
 Sets the color values in the blue channel.
 This is a wrapper method that calls
 \code{setChannel(obj, channel="blue", ...)}.
}

\arguments{
 \item{...}{Arguments accepted by \code{\link[R.colors:setChannel.RgbColor]{*setChannel}()}.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{color}

\seealso{
  For more information see \code{\link{RgbColor}}.
}
\keyword{internal}
\keyword{methods}
