%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RgbColor.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setGreen.RgbColor}
\alias{setGreen.RgbColor}
\alias{RgbColor.setGreen}
\alias{setGreen.RgbColor}
\alias{setGreen,RgbColor-method}


\title{Sets the color values in the green channel}

\usage{\method{setGreen}{RgbColor}(this, ...)}

\description{
 Sets the color values in the green channel.
 This is a wrapper method that calls
 \code{setChannel(obj, channel="green", ...)}.
}

\arguments{
 \item{...}{Arguments accepted by \code{\link[R.colors:setChannel.RgbColor]{*setChannel}()}.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{color}

\seealso{
  For more information see \code{\link{RgbColor}}.
}
\keyword{internal}
\keyword{methods}
