%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RgbColor.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setRed.RgbColor}
\alias{setRed.RgbColor}
\alias{RgbColor.setRed}
\alias{setRed.RgbColor}
\alias{setRed,RgbColor-method}


\title{Sets the color values in the red channel}

\usage{\method{setRed}{RgbColor}(this, ...)}

\description{
 Sets the color values in the red channel.
 This is a wrapper method that calls
 \code{setChannel(obj, channel="red", ...)}.
}

\arguments{
 \item{...}{Arguments accepted by \code{\link[R.colors:setChannel.RgbColor]{*setChannel}()}.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{color}

\seealso{
  For more information see \code{\link{RgbColor}}.
}
\keyword{internal}
\keyword{methods}
