%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  VgaColor.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{VgaColor}
\docType{class}
\alias{VgaColor}


\title{Class representing (categorial) colors in VGA}

\description{
 Package:  R.colors \cr
\bold{Class VgaColor}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[R.colors]{Color}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{VgaColor}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{VgaColor}\cr
extends \emph{\link[R.colors]{Color}}\cr



 Class representing (categorial) colors in VGA.
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{getColors} \tab  -\cr
 \tab \code{getColorSpace} \tab  -\cr
}


 \bold{Methods inherited from Color}:\cr
[, as.character, display, drawColorRamp, getColorFilter, getColors, getColorSpace, length, palette, plotParallel, seq, setColorFilter, viewThroughColorFilter

 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


}

\usage{VgaColor()}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{color}

\references{
  [1] F. David McRitchie,
      \emph{Color Palette and the 56 Excel ColorIndex Colors}, 2005.
      \url{http://www.mvps.org/dmcritchie/excel/colors.htm}
}

\seealso{
  \code{\link[grDevices]{palette}}.
}
\keyword{classes}
