%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Color.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{viewThroughColorFilter.Color}
\alias{viewThroughColorFilter.Color}
\alias{Color.viewThroughColorFilter}
\alias{viewThroughColorFilter.Color}
\alias{viewThroughColorFilter,Color-method}


 \title{Filters a set of colors using the default filter}

 \synopsis{viewThroughColorFilter.Color(this, colors, ...)}
\usage{Color$viewThroughColorFilter(colors, ...)}

 \description{
  Filters a set of colors using the default filter.

  To the developer of a Color class:
  \emph{This method is to be implemented in a subclass.}
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \keyword{color}

 \seealso{
   To set or reset the default filter see \code{\link[R.colors:setColorFilter.Color]{*setColorFilter}()}.
   To get the default filter see \code{\link[R.colors:getColorFilter.Color]{*getColorFilter}()}.
   For more information see \code{\link{Color}}.
 }
\keyword{internal}
\keyword{methods}
